Move field availability unit tests to separate file
